requirements.txt
httpcore==1.0.9
httpx==0.28.1
pluggy==1.6.0
pytest==9.0.2
pytest-html==4.1.1
pytest-metadata==3.1.1
PyYAML==6.0.3
requests==2.32.5
urllib3==2.6.2